Improve math, improve renderer
