+++ egolib/src/egolib/Extensions/ogl_extensions.c
@@ -526,9 +526,3 @@ void oglx_begin_culling( GLenum face, GLenum mode )
 }
 
 //--------------------------------------------------------------------------------------------
-void oglx_end_culling( void )
-{
-    GL_DEBUG(glDisable)(GL_CULL_FACE);  // GL_ENABLE_BIT
-}
-
-//--------------------------------------------------------------------------------------------
