+++ egolib/src/egolib/Extensions/ogl_extensions.h
@@ -227,4 +227,3 @@ struct oglx_culling_data_t
 };
 
 void oglx_begin_culling(GLenum which, GLenum mode );
-void oglx_end_culling();
