+++ egolib/src/egolib/Math/TemplateUtilities.hpp
@@ -135,3 +135,46 @@ struct make_index_sequence<0, S...> : public index_sequence<S ...>
 
 } // namespace Core
 } // namespace Ego
+
+namespace Ego {
+namespace Math {
+namespace Internal {
+
+template <size_t I, typename Type, size_t Size>
+void unpack(Type(&dst)[Size]);
+
+template <size_t I, typename  Type, size_t Size, typename Arg>
+void _unpack(Type(&dst)[Size], Arg&& arg) {
+	dst[I] = arg;
+}
+
+template <size_t I, typename Type, size_t Size, typename Arg, class ... Args>
+void _unpack(Type(&dst)[Size], Arg&& arg, Args&& ... args) {
+	dst[I] = arg;
+	_unpack<I + 1, Type>(dst, args ...);
+}
+
+/**
+* @brief
+*  You can use "unpack" for storing non-type parameter packs in arrays.
+* @invariant
+*  The number of arguments must be exactly the length of the array.
+* @invariant
+*  The argument types must be the array type.
+* @remark
+*  Example usage:
+*  @code
+*  float a[3]; unpack(a,0.5,0.1)
+*  @endcode
+* @author
+*  Michael Heilmann
+*/
+template <typename Type, size_t Size, typename ... Args>
+void unpack(Type(&dst)[Size], Args&& ...args) {
+	static_assert(Size == sizeof ... (args), "wrong number of arguments");
+	_unpack<0, Type, Size>(dst, args ...);
+}
+
+} // namespace Internal
+} // namespace Math
+} // namespace Ego
